add authentication middleware

+++ backend/controllers/taskController.js
@@ -1,9 +1,10 @@
 const asyncHandler = require('express-async-handler');
 const Task = require('../models/taskmodel')
+const User = require('../models/userModel')
 
 
 const getTasks = asyncHandler(async(req, res) =>{
-    const tasks = await Task.find()
+    const tasks = await Task.find({user: req.user.id})
     res.status(200).json(tasks)
 })
 
@@ -13,7 +14,7 @@ const setTasks = asyncHandler(async(req, res) =>{
         throw new Error('Please enter a task');
     }
 
-    const task = await Task.create({ text: req.body.text})
+    const task = await Task.create({ text: req.body.text, user: req.user.id})
     res.status(200).json(task);
 })
 
@@ -22,6 +23,16 @@ const updateTask = asyncHandler(async(req, res) => {
     if (!task) {
         res.status(400);
         throw new Error("Task not found");
+    }
+    
+    const user = await User.findById(req.user.id);
+    if( !user){
+        res.status(401)
+        throw new Error("no user found");
+    }
+    if(task.user.toString() !== user.id){
+        res.status(401)
+        throw new Error("user not authorized");
     }
 
     const updatedTask = await Task.findByIdAndUpdate(req.params.id, req.body, { new: true });
@@ -35,6 +46,15 @@ const deleteTask = asyncHandler(async(req, res) => {
         throw new Error("Task not found");
     }
 
+    const user = await User.findById(req.user.id);
+    if( !user){
+        res.status(401)
+        throw new Error("no user found");
+    }
+    if(task.user.toString() !== user.id){
+        res.status(401)
+        throw new Error("user not authorized to delete");
+    }
     const deletedTask = await Task.findByIdAndDelete(req.params.id);
     res.status(200).json(deletedTask);
 });
